Move parser code to psl-parser package
Moved all the parser code to the psl-parser package, and removed some of
the dependencies on other modules that should not be part of the parser
package.

Cleaned up all the npm package related files from the `src/parser`
directory, as this directory does not contain an isolated npm package
anymore.